FarmConnect/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.0",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-crispy-forms>=2.5",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt",
    "dj-rest-auth[with_social]",
    "drf-spectacular>=0.29.0",
    "faker>=38.2.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]
